refactoring: 코드 정리
