+++ index.html
@@ -14,9 +14,10 @@
       </div>
       <ul id="todoList"></ul>
       <!-- 필터 테스트하려고 만든 임의의 버튼  -->
-      <button id = 'leftTodo'>모두</button>
-      <button id = 'centerTodo'>남은거</button>
-      <button id = 'rightTodo'>완성한 거</button>
+      <button id = 'allTodo'>모두</button>
+      <button id = 'leftTodo'>남은거</button>
+      <button id = 'completedTodo'>완성한 거</button>
+      <button id = 'completedAll'>모두 완성</button>
     </div>
     <script type="module" src="index.js"></script>
   </body>
